Clean up a few things
